python_backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0

# Document processing
python-pptx==0.6.23
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8

# Video/Audio processing
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3

# Optional TTS (can be added later)
# TTS==0.20.2
# torch==2.1.0
# torchaudio==2.1.0
